SWATCH-2460: Move the /remittance/purge endpoint to the billing API
Move the endpoint "/internal/rpc/tally/remittance/purge" from internal tally API to the internal billing API. The new path should be "/internal/rpc/remittance/purge" instead.
